ExpenseDocAI/src/docs/tests/requirements.txt
pytest==8.3.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-metadata==3.0.0
pytest-mock==3.14.0
pytest-benchmark==5.1.0
requests==2.31.0
python-dotenv==1.0.1 